+++ docs/include/lib/file_storage.inc.php
@@ -10,12 +10,43 @@
 /* modify it under the terms of the GNU General Public License			*/
 /* as published by the Free Software Foundation.						*/
 /************************************************************************/
-// $Id: constants.inc.php 5929 2006-03-06 16:37:08Z cridpath $
+// $Id$
 
 if (!defined('AT_INCLUDE_PATH')) { exit; }
 
-function fs_authenticate($owner_type, $owner_id) {
+define('WORKSPACE_AUTH_NONE',  0);
+define('WORKSPACE_AUTH_READ',  1);
+define('WORKSPACE_AUTH_WRITE', 2); 
+define('WORKSPACE_AUTH_RW',    3); // to save time
 
+// given an owner_type and owner_id
+// returns false if user cannot read or write to this workspace
+// returns WORKSPACE_AUTH_READ if the user can read
+// returns WORKSPACE_AUTH_WRITE if the user can write
+function fs_authenticate($owner_type, $owner_id) {
+	if ($owner_type == WORKSPACE_PERSONAL) {
+		if ($owner_id == $_SESSION['member_id']) {
+			return WORKSPACE_AUTH_RW;
+		}
+		return WORKSPACE_AUTH_NONE;
+	}
+	if ($owner_type == WORKSPACE_ASSIGNMENT) {
+
+	}
+	if ($owner_type == WORKSPACE_GROUP) {
+		if (isset($_SESSION['groups'][$owner_id])) {
+			return WORKSPACE_AUTH_RW;
+		}
+		return WORKSPACE_AUTH_NONE;
+	}
+	if ($owner_type == WORKSPACE_COURSE) {
+		if (authenticate(AT_PRIV_FILE_STORAGE, AT_PRIV_RETURN)) {
+			return WORKSPACE_AUTH_RW;
+		}
+		return WORKSPACE_AUTH_READ;
+	}
+
+	return WORKSPACE_AUTH_NONE;
 }
 
 /**
@@ -66,14 +97,15 @@ function fs_print_folders($current_folder_id, $parent_folder_id, &$folders, $dis
  * $file_id ID of a file in a revision sequence. can be any revision, does not have to be the latest.
  * This function is recursive and uses fs_get_revisions_down_recursive() and fs_get_revisions_recurisve() below.
  */
-function fs_get_revisions($file_id) {
+function fs_get_revisions($file_id, $owner_type, $owner_id) {
 	global $db;
 
 	$sql = "SELECT * FROM ".TABLE_PREFIX."files WHERE file_id=$file_id";
 	$result = mysql_query($sql, $db);
-	$row = mysql_fetch_assoc($result);
-
-	return array_merge(array_reverse(fs_get_revisions_down_recursive($row['parent_file_id'])), array($row), fs_get_revisions_recursive($file_id));
+	if ($row = mysql_fetch_assoc($result)) {
+		return array_merge(array_reverse(fs_get_revisions_down_recursive($row['parent_file_id'])), array($row), fs_get_revisions_recursive($file_id));
+	}
+	return array();
 }
 
 /**
@@ -154,9 +186,9 @@ function fs_get_file_path($file_id) {
  *
  * $file_id the ID of the file to delete. can be any ID within a revision sequence.
  */
-function fs_delete_file($file_id) {
+function fs_delete_file($file_id, $owner_type, $owner_id) {
 	global $db;
-	$revisions = fs_get_revisions($file_id);
+	$revisions = fs_get_revisions($file_id, $owner_type, $owner_id);
 	foreach ($revisions as $file) {
 		$sql = "DELETE FROM ".TABLE_PREFIX."files WHERE file_id=$file[file_id]";
 		mysql_query($sql, $db);
@@ -204,26 +236,26 @@ function fs_get_file_type_icon($file_name) {
  *
  * $folder_id the ID of the folder to delete, recursively, with content.
  */
-function fs_delete_folder($folder_id) {
+function fs_delete_folder($folder_id, $owner_type, $owner_id) {
 	if (!$folder_id) { return; }
 
 	global $db;
 
-	$sql = "SELECT folder_id FROM ".TABLE_PREFIX."folders WHERE parent_folder_id=$folder_id";
+	$sql = "SELECT folder_id FROM ".TABLE_PREFIX."folders WHERE parent_folder_id=$folder_id AND owner_type=$owner_type AND owner_id=$owner_id";
 	$result = mysql_query($sql, $db);
 	while ($row = mysql_fetch_assoc($result)) {
-		fs_delete_folder($row['folder_id']);
+		fs_delete_folder($row['folder_id'], $owner_type, $owner_id);
 	}
 
-	$sql = "DELETE FROM ".TABLE_PREFIX."folders WHERE folder_id=$folder_id";
+	$sql = "DELETE FROM ".TABLE_PREFIX."folders WHERE folder_id=$folder_id AND owner_type=$owner_type AND owner_id=$owner_id";
 	mysql_query($sql, $db);
 
 	// delete this file's folders (we only select the latest versions because
 	// the delete_file() function takes care of the revisions for us
-	$sql = "SELECT file_id FROM ".TABLE_PREFIX."files WHERE folder_id=$folder_id AND parent_file_id=0";
+	$sql = "SELECT file_id FROM ".TABLE_PREFIX."files WHERE folder_id=$folder_id AND parent_file_id=0 AND owner_type=$owner_type AND owner_id=$owner_id";
 	$result = mysql_query($sql, $db);
 	while ($row = mysql_fetch_assoc($result)) {
-		fs_delete_file($row['file_id']);
+		fs_delete_file($row['file_id'], $owner_type, $owner_id);
 	}
 }
 
@@ -234,15 +266,15 @@ function fs_delete_folder($folder_id) {
  * $zipfile reference to the zipFile object.
  * $path the absolute path to the current folder.
  */
-function fs_download_folder($folder_id, &$zipfile, $path = '') {
+function fs_download_folder($folder_id, &$zipfile, $owner_type, $owner_id, $path = '') {
 	global $db;
-	$sql = "SELECT title FROM ".TABLE_PREFIX."folders WHERE folder_id=$folder_id";
+	$sql = "SELECT title FROM ".TABLE_PREFIX."folders WHERE folder_id=$folder_id AND owner_type=$owner_type AND owner_id=$owner_id";
 	$result = mysql_query($sql, $db);
 	$parent_row = mysql_fetch_assoc($result);
 
 	$zipfile->create_dir($path . $parent_row['title']);
 
-	$sql = "SELECT file_id, file_name, UNIX_TIMESTAMP(date) AS date FROM ".TABLE_PREFIX."files WHERE folder_id=$folder_id AND parent_file_id=0";
+	$sql = "SELECT file_id, file_name, UNIX_TIMESTAMP(date) AS date FROM ".TABLE_PREFIX."files WHERE folder_id=$folder_id AND parent_file_id=0 AND owner_type=$owner_type AND owner_id=$owner_id";
 	$result = mysql_query($sql, $db);
 	while ($row = mysql_fetch_assoc($result)) {
 		$file_path = fs_get_file_path($row['file_id']) . $row['file_id'];
@@ -250,10 +282,10 @@ function fs_download_folder($folder_id, &$zipfile, $path = '') {
 		$zipfile->add_file(file_get_contents($file_path), $path . $parent_row['title'] .'/' . $row['file_name'], $row['date']);
 	}
 
-	$sql = "SELECT folder_id FROM ".TABLE_PREFIX."folders WHERE parent_folder_id=$folder_id";
+	$sql = "SELECT folder_id FROM ".TABLE_PREFIX."folders WHERE parent_folder_id=$folder_id AND owner_type=$owner_type AND owner_id=$owner_id";
 	$result = mysql_query($sql, $db);
 	while ($row = mysql_fetch_assoc($result)) {
-		fs_download_folder($row['folder_id'], $zipfile, $path . $parent_row['title'] . '/');
+		fs_download_folder($row['folder_id'], $zipfile, $owner_type, $owner_id, $path . $parent_row['title'] . '/');
 	}
 }
 
@@ -264,8 +296,8 @@ function fs_download_folder($folder_id, &$zipfile, $path = '') {
  * $workspace the owner_type of this folder
  * $owner_id the ID of the owner.
  */
-function fs_get_folder_path($folder_id, $workspace, $owner_id) {
-	$folder_path = fs_get_folder_path_recursive($folder_id, $workspace, $owner_id);
+function fs_get_folder_path($folder_id, $owner_type, $owner_id) {
+	$folder_path = fs_get_folder_path_recursive($folder_id, $owner_type, $owner_id);
 
 	return array_reverse($folder_path);
 }
@@ -274,18 +306,18 @@ function fs_get_folder_path($folder_id, $workspace, $owner_id) {
  * recursively return the path to the current folder
  * PRIVATE! do not call directly, use get_folder_path() above.
  */
-function fs_get_folder_path_recursive($folder_id, $workspace, $owner_id) {
+function fs_get_folder_path_recursive($folder_id, $owner_type, $owner_id) {
 	global $db;
 
 	if ($folder_id == 0) {
 		return array();
 	}
 
-	$sql = "SELECT folder_id, title, parent_folder_id FROM ".TABLE_PREFIX."folders WHERE folder_id=$folder_id AND owner_type=$workspace AND owner_id=$owner_id";
+	$sql = "SELECT folder_id, title, parent_folder_id FROM ".TABLE_PREFIX."folders WHERE folder_id=$folder_id AND owner_type=$owner_type AND owner_id=$owner_id";
 	$result = mysql_query($sql, $db);
 	$row = mysql_fetch_assoc($result);
 
-	return array_merge(array($row), fs_get_folder_path_recursive($row['parent_folder_id'], $workspace, $owner_id));
+	return array_merge(array($row), fs_get_folder_path_recursive($row['parent_folder_id'], $owner_type, $owner_id));
 }
 
 ?>
